start server from jest
