+++ src/tests/register.test.ts
@@ -1,11 +1,14 @@
 import { request } from "graphql-request";
 
-import { host } from "./constants";
 import { User } from "../entity/User";
-import { createTypeormConn } from "../utils/createTypeormConn";
+import { startServer } from "../startServer";
+
+let getHost = () => "";
 
 beforeAll(async () => {
-  await createTypeormConn();
+  const app = await startServer();
+  const { port } = app.address();
+  getHost = () => `http://127.0.0.1:${port}`;
 });
 
 const email = "tom@bob.com";
@@ -18,7 +21,7 @@ mutation {
 `;
 
 test("Register user", async () => {
-  const response = await request(host, mutation);
+  const response = await request(getHost(), mutation);
   expect(response).toEqual({ register: true });
   const users = await User.find({ where: { email } });
   expect(users).toHaveLength(1);
@@ -26,7 +29,3 @@ test("Register user", async () => {
   expect(user.email).toEqual(email);
   expect(user.password).not.toEqual(password);
 });
-
-// use a test database
-// drop all data once the test is over
-// when I run yarn test it also starts the server
